requirements.txt
streamlit==1.52.1
streamlit-chat==0.1.1

openai==2.11.0
groq==0.37.1
python-dotenv==1.2.1

langchain==1.1.3
langchain-core==1.1.3
langchain-community==0.4.1
langchain-huggingface==1.1.0
langchain-text-splitters==1.0.0

sentence-transformers==5.2.0
transformers==4.57.3
semantic-text-splitter==0.28.0

qdrant-client==1.7.0
pypdf==6.4.1
beautifulsoup4==4.14.3

numpy==2.3.5
pandas==2.3.3
scikit-learn==1.8.0
scipy==1.16.3

requests==2.32.5
pyyaml==6.0.3
regex==2025.11.3
tqdm==4.67.1
pyarrow==22.0.0
